requirements.txt
langchain
langchain-openai
langgraph
fastapi
uvicorn
streamlit
python-dotenv
faiss-cpu
tiktoken
pydantic
langchain-community
langchain-text-splitters